Extracted qdrant_manager and services into separate locations
